% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{groups}
\alias{groups}
\title{All self-determination movements}
\format{
A data frame with 5549 rows and 5 variables:
\describe{
  \item{kgcid}{group's unique ID}
  \item{group_name}{group's name}
  \item{year}{calendar year}
  \item{country}{country the group is challenging}
  \item{country_iso3c}{the ISO3C for the country the group is challenging}
}
}
\source{
\url{http://www.kathleengallaghercunningham.com/data-projects.html}
}
\usage{
groups
}
\description{
A dataset containing details of all self-determination movements and the
countries they challenged from 1960 to 2005.
}
\keyword{datasets}
